Setup Basic Framework for Menus
I create the main menu and option menu scenes and created the UI for the
menus. I made a function to load unity scenes. Currently the start game,
options, main menu btns all function and go to there respective scenes.
I also fixed the camera on the test level scene to use the smoothfollow
script, and I added a player script for noah to reference