Updated Menus
I updated the Option menu to now have a credits canvas which is shown
when the credits btn is pressed. I also added a placeholder button for
configuring controller setting later down the line. I have also used
some of the background art assets to decorate the options and main menu
scenes. In main menu scene I added code for the exit button that should
cause the apllication to quit when pressed.