Polished the Game for Milestone 1 + fix controller bug
Here we go, I fixed the controller bug that was causing both players to
jump when the jump button was pressed on either controller. I made and
added a Crate art asset to replace the unity pink crate, and made a
crate prefab. The Destroy function for the player now works for this
build, but will need adjustments for down the line. We now have a
killzone prefab that if the player foes into that area they will die.
Finally I went through the test Level and polished up the level for
playtesting. Overall lot of stuff done and the game is looking good.
:+1: