Character Selection Has been implemented
I have gotten the character selection up and running. The Drop in drop
out feature has not been implemented yet. To submit on P1 press the
space button, to submit on P2 press the enter button. I will be adding
in instruction icons and the drop in/ drop out feature in the next
commit.