Added Goal Prefab
I made a goal prefab to move the players to next level when it is
touched. Fun fact the goal is the Tardis( iwill replace it with a better
quality image later). I changed the level 1 scene name from "jason's
test scene" to "level 1". I also added level 1 to build. I still need to
implement the code for setting up characters in level 1