Polishing Game Part 2
Ok so a lot of stuff has been changed for polish. I replace the bigArrow
art asset to be a big white arrow so that it would be easier to see for
players. I made a bullet asset for enemies and replace the crate asset
with the bullet asset in the bullet prefab. I added pulp/noir comic
cover for use as collectables and made prefabs for the collectables.
Note there is currently a bug in the obstacle script that is causing
collectables to fall through the world. I will fix this bug in next
commit. I made the UI scale depending on resolution size and screen
size. We  now have a Noir font that fits the game called bettyNoir and
it is open source. I have gone through all the UI and Text in the Game
and Replaced it with this font. I also adjust the highlight button color
to from neon green to a nice dark cream brown.