Polishing game part 3
The next wave of polish updates is upon us.  With Jason's help we fixed
the collectible bug, and added collectibles into both level 1 and test
level.  We  also had a mini playtest with Josh and Robert which brought
up some of the following issues which have been fixed.  I added some
changes in enviorment assets in level 1 and Test to fix some issuse with
detecting foreground and background. I also changed the signs in the
level to be more clear to players. I added the fence asset that jason
made into level 1 to prevent player from going places. In the sewer of
level 1 we added a co-op challenge.