Made and implemented Game UI
I made and implement the game UI. the green bar is your health and the
purple bar is  the pulp power bar. When a player dies (if there are 2
players) their character profile sprite will be replace with a version
of an x over it. It will change the player profile depending on which
character the player is and whether or not there is 2 players. I will be
adding more features to it soon.