Changed color of Fire Escapes, and added a little code to enemies
I changed the color of the fire escapes to black so that their easier to
see against buildings, they are hard to see against the paralax layer
though so I might have to change them, I tried a light blue color, but
they looked wierd.

I added some code to the enemies so that when they face the other
direction they should turn but it doesn't seem to work will continue
working on it.