I fixed the bug wall clinging bug
I fixed the bug where if player was providing input againist an obstical
in the air it would cause the player to not fall down the wall. but
instead remain in that position. I do not know exactly what is causing
this bug but i am guessing it has to do with back-end unity stuff. The
way i was able to fix it was by creating a bool called aircontrol that
gives the player the ability to manuver in the air. If the player
colided with anything in the air, they would lose their ability to
manuver in the air until they are no longer colliding with it.