Setup Tutorial for the game
I setup the tutorial level for the game so that it can be played. I
added it to the build and reconfigured the level order. I also added in
the necessary data need for it to run. I fixed the direction of the
enemies movements. Added game UI to tutorial level.