Fixed some UI Features + Added View controls to options
I added the ability to go back to the main menu scene from the character
selection by pressing the pause button.  In the option Scene I added the
ability to view player controls from the option menu and you can select
which type of controls you want to see. I also have fix the bug where
the event system would  not change current selected based on which menu
is up. It now will change the current selected so that the controller
will never lose a selected button.